ml/codes/hands-on-ml/pyproject.toml
[project]
name = "hands-on-ml"
version = "0.1.0"
description = ""
authors = [{ name = "jargonzhou", email = "zhoujiagen@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "scikit-learn (>=1.7.1,<2.0.0)",
    "keras (>=3.11.3,<4.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "tensorflow (>=2.20.0,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
]

[tool.poetry]
packages = [{ include = "hands_on_ml", from = "src" }]

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.3.2"
pylint = "^3.3.8"
pandas-stubs = "^2.3.2.250827"
tensorflow-datasets = "^4.9.9"
pydot = "^4.0.1"
tensorboard-plugin-profile = "^2.20.6"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
